+++ Changes.rst
@@ -5,6 +5,9 @@ Planned changes
 Possible changes
 * add more info to qname to make BAM analysis faster/easier
 
+2015.10.05
+* Read length information added to qname (1.16.0.dev0)
+
 2015.10.01
 * Instead of making several different files write out the alignment accuracy in the original BAM itself.
   Still produce a perfect BAM as needed